Fixes in response to PR feedback.
