+++ test/suite/test_timestamp29.py
@@ -31,14 +31,10 @@
 
 import time
 import wiredtiger, wttest
-from helper_disagg import disagg_test_class
 from wiredtiger import stat
 
 # Timestamps: Test setting and querying the stable disaggregated schema epoch.
-@disagg_test_class
 class test_timestamp29(wttest.WiredTigerTestCase):
-
-    conn_config = 'statistics=(all),disaggregated=(role="leader",lose_all_my_data=true)'
 
     def get_stat(self, stat_name):
         stat_cursor = self.session.open_cursor('statistics:', None, None)
@@ -49,11 +45,6 @@ class test_timestamp29(wttest.WiredTigerTestCase):
     def assertEpochEqual(self, expected_ts):
         self.assertTimestampsEqual(
             self.conn.query_timestamp('get=stable_disaggregated_schema_epoch'),
-            self.timestamp_str(expected_ts))
-
-    def assertLastCheckpointEpochEqual(self, expected_ts):
-        self.assertTimestampsEqual(
-            self.conn.query_timestamp('get=last_disaggregated_schema_epoch'),
             self.timestamp_str(expected_ts))
 
     def assertStatEqual(self, stat_name, expected_value, retries=10):
@@ -154,58 +145,3 @@ class test_timestamp29(wttest.WiredTigerTestCase):
         # Stats: eight calls, five updates.
         self.assertStatEqual(stat.conn.txn_set_ts_stable_disagg_epoch, 8)
         self.assertStatEqual(stat.conn.txn_set_ts_stable_disagg_epoch_upd, 5)
-
-    def test_checkpoint(self):
-        '''
-        Test that the stable disaggregated schema epoch is included in the checkpoint, that it is
-        correctly picked up on restart, and that it can be updated in a checkpoint.
-        '''
-        # Create a layered table so checkpoints write through disaggregated storage.
-        uri = 'layered:test_timestamp29'
-        self.session.create(uri, 'key_format=S,value_format=S')
-
-        # The first checkpoint, taken before the epoch is set, records 0 in the checkpoint.
-        self.session.checkpoint()
-        self.assertLastCheckpointEpochEqual(0)
-
-        # Set the timestamps and the epoch, checkpoint, and verify that the epoch value is included
-        # in the checkpoint.
-        self.conn.set_timestamp(
-            'oldest_timestamp=' + self.timestamp_str(1) +
-            ',stable_timestamp=' + self.timestamp_str(10) +
-            ',stable_disaggregated_schema_epoch=' + self.timestamp_str(10))
-        self.assertEpochEqual(10)
-        self.session.checkpoint()
-        self.assertLastCheckpointEpochEqual(10)
-
-        # Restart the connection. The new connection should pick up the latest epoch value from the
-        # latest checkpoint.
-        with self.expectedStdoutPattern('Removing local file'):
-            self.reopen_conn()
-        self.assertLastCheckpointEpochEqual(10)
-
-        # Advancing the epoch and the timestamps should update the epoch in the checkpoint to the
-        # new value.
-        self.conn.set_timestamp(
-            'oldest_timestamp=' + self.timestamp_str(20) +
-            ',stable_timestamp=' + self.timestamp_str(30) +
-            ',stable_disaggregated_schema_epoch=' + self.timestamp_str(30))
-        self.assertEpochEqual(30)
-        self.session.checkpoint()
-        self.assertLastCheckpointEpochEqual(30)
-
-        # A checkpoint without changing the epoch leaves it at the same value.
-        self.conn.set_timestamp('stable_timestamp=' + self.timestamp_str(40))
-        self.session.checkpoint()
-        self.assertLastCheckpointEpochEqual(30)
-
-        # A checkpoint that changes only the epoch also advances the epoch value in the checkpoint.
-        self.conn.set_timestamp('stable_disaggregated_schema_epoch=' + self.timestamp_str(40))
-        self.session.checkpoint()
-        self.assertLastCheckpointEpochEqual(40)
-
-        # Restart the connection to check that the new value was preserved.
-        with self.expectedStdoutPattern('Removing local file'):
-            self.reopen_conn()
-        self.assertEpochEqual(0) # The "current" epoch value must be set explicitly by the caller.
-        self.assertLastCheckpointEpochEqual(40)
